requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django==3.2
django-admin
pandas
pytest==6.2.0
pytest-django
psycopg2-binary
djangorestframework==3.11.1
djangorestframework-simplejwt
django-import-export
loguru

